Start moving shared code to BaseTextAttachmentMessageCell
